requirements.txt
pandas==1.5.3
numpy==1.24.4
scikit-learn==1.2.2
xgboost==1.7.6
joblib==1.2.0
boto3==1.28.0
s3fs>=2023.6.0
python-dotenv==1.0.0
pyyaml
# -e .